pyproject.toml
[tool.poetry]
name = "chatroom"
version = "0.1.0"
description = ""
authors = ["Jeffrey04 <mycyberpet@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
channels = "^2"
channels-redis = "^3.2.0"
django-notifications-hq = "^1.6.0"
django-environ = "^0.4.5"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
prospector = "^1.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
